GUI progress 90%

+++ src/main/java/it/polimi/ingsw/Client/GUI/Controllers/Point.java
@@ -1,19 +1,19 @@
 package it.polimi.ingsw.Client.GUI.Controllers;
 
 public class Point {
-    private int x, y;
+    private double x, y;
 
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
